Title: 합과 평균, Time: 6ms, Memory: 0MB, Status: Passed - CodeTree

+++ 240119/합과 평균/sum-and-avg.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 using namespace std;
 int main() {
-    double a, b;
+    int a, b;
     cin >> a >> b;
     cout << fixed;
     cout.precision(1);
